Create GEOS Geometry Vectors

Create GEOS Geometry Vectors

# S3 method for wk_xy
as_geos_geometry(x, ...)

# S3 method for wk_xyz
as_geos_geometry(x, ...)

as_geos_geometry(x, ...)

# S3 method for geos_geometry
as_geos_geometry(x, ...)

# S3 method for default
as_geos_geometry(x, ...)

# S3 method for character
as_geos_geometry(x, ..., crs = NULL)

# S3 method for blob
as_geos_geometry(x, ..., crs = NULL)

# S3 method for WKB
as_geos_geometry(x, ..., crs = NULL)

geos_geometry(crs = wk::wk_crs_inherit())

Arguments

x

An object to be coerced to a geometry vector

...

Unused

crs

An object that can be interpreted as a CRS. See wk::wk_crs().

Value

A geos geometry vector

Examples

as_geos_geometry("LINESTRING (0 1, 3 9)")
#> <geos_geometry[1]>
#> [1] <LINESTRING (0 1, 3 9)>