Buffer a geometry

geos_buffer() returns a polygon or multipolygon geometry.

geos_offset_curve() returns a linestring offset to the left by distance.

geos_buffer(geom, distance, params = geos_buffer_params())

geos_offset_curve(geom, distance, params = geos_buffer_params())

geos_buffer_params(
  quad_segs = 30,
  end_cap_style = c("round", "flat", "square"),
  join_style = c("round", "mitre", "bevel"),
  mitre_limit = 1,
  single_sided = FALSE
)

Arguments

geom
A GEOS geometry vector

distance
The buffer distance. Can be negative to buffer or offset on the righthand side of the geometry.

params
A geos_buffer_params()

quad_segs
The number of segments per quadrant. A higher number here will increase the apparent resolution of the resulting polygon.

end_cap_style
One of "round", "flat", or "square".

join_style
One of "round", "mitre", or "bevel".

mitre_limit
If join_style is "mitre", the relative extent (from zero to one) of the join.

single_sided
Use TRUE to buffer on only the right side of the geometry. This does not apply to geos_offset_curve(), which is always one-sided.

Value

A GEOS geometry vector along the recycled length of geom and distance.

Examples

geos_buffer("POINT (0 0)", 1)

#> <geos_geometry[1]>
#> [1] POLYGON ((1 0, 0.99863 -0.052336, 0.99452 -0.10453, 0.98769 -0.15643, 0.97815 -0.20791, 0.96593 -0.25882, 0.95106 -0.30902, 0.93358 -0.35837, 0.91355 -0.40674, 0.89101 -0.45399, 0.86603 -0.5, 0.83867 -0.54464, 0.80902 -0.58779, 0.77715 -0.62932, 0.74314 -0.66913, 0.70711 -0.70711, 0.66913 -0.74314, 0.62932 -0.77715, 0.58779 -0.80902, 0.54464 -0.83867, 0.5 -0.86603, 0.45399 -0.89101, 0.40674 -0.91355, 0.35837 -0.93358, 0.30902 -0.95106, 0.25882 -0.96593, 0.20791 -0.97815, 0.15643 -0.98769, 0.10453 -0.99452, 0.052336 -0.99863, 2.5037e-15 -1, -0.052336 -0.99863, -0.10453 -0.99452, -0.15643 -0.98769, -0.20791 -0.97815, -0.25882 -0.96593, -0.30902 -0.95106, -0.35837 -0.93358, -0.40674 -0.91355, -0.45399 -0.89101, -0.5 -0.86603, -0.54464 -0.83867, -0.58779 -0.80902, -0.62932 -0.77715, -0.66913 -0.74314, -0.70711 -0.70711, -0.74314 -0.66913, -0.77715 -0.62932, -0.80902 -0.58779, -0.83867 -0.54464, -0.86603 -0.5, -0.89101 -0.45399, -0.91355 -0.40674, -0.93358 -0.35837, -0.95106 -0.30902, -0.96593 -0.25882, -0.97815 -0.20791, -0.98769 -0.15643, -0.99452 -0.10453, -0.99863 -0.052336, -1 -6.7838e-15, -0.99863 0.052336, -0.99452 0.10453, -0.98769 0.15643, -0.97815 0.20791, -0.96593 0.25882, -0.95106 0.30902, -0.93358 0.35837, -0.91355 0.40674, -0.89101 0.45399, -0.86603 0.5, -0.83867 0.54464, -0.80902 0.58779, -0.77715 0.62932, -0.74314 0.66913, -0.70711 0.70711, -0.66913 0.74314, -0.62932 0.77715, -0.58779 0.80902, -0.54464 0.83867, -0.5 0.86603, -0.45399 0.89101, -0.40674 0.91355, -0.35837 0.93358, -0.30902 0.95106, -0.25882 0.96593, -0.20791 0.97815, -0.15643 0.98769, -0.10453 0.99452, -0.052336 0.99863, -5.5128e-15 1, 0.052336 0.99863, 0.10453 0.99452, 0.15643 0.98769, 0.20791 0.97815, 0.25882 0.96593, 0.30902 0.95106, 0.35837 0.93358, 0.40674 0.91355, 0.45399 0.89101, 0.5 0.86603, 0.54464 0.83867, 0.58779 0.80902, 0.62932 0.77715, 0.66913 0.74314, 0.70711 0.70711, 0.74314 0.66913, 0.77715 0.62932, 0.80902 0.58779, 0.83867 0.54464, 0.86603 0.5, 0.89101 0.45399, 0.91355 0.40674, 0.93358 0.35837, 0.95106 0.30902, 0.96593 0.25882, 0.97815 0.20791, 0.98769 0.15643, 0.99452 0.10453, 0.99863 0.052336, 1 0))

geos_offset_curve("LINESTRING (0 0, 0 10, 10 0)", 1)

#> <geos_geometry[1]>
#> [1] LINESTRING (-1 0, -1 10, -0.99863 10.052, -0.99452 10.105, -0.98769 10.156, -0.97815 10.208, -0.96593 10.259, -0.95106 10.309, -0.93358 10.358, -0.91355 10.407, -0.89101 10.454, -0.86603 10.5, -0.83867 10.545, -0.80902 10.588, -0.77715 10.629, -0.74314 10.669, -0.70711 10.707, -0.66913 10.743, -0.62932 10.777, -0.58779 10.809, -0.54464 10.839, -0.5 10.866, -0.45399 10.891, -0.40674 10.914, -0.35837 10.934, -0.30902 10.951, -0.25882 10.966, -0.20791 10.978, -0.15643 10.988, -0.10453 10.995, -0.052336 10.999, 2.8328e-16 11, 0.052336 10.999, 0.10453 10.995, 0.15643 10.988, 0.20791 10.978, 0.25882 10.966, 0.30902 10.951, 0.35837 10.934, 0.40674 10.914, 0.45399 10.891, 0.5 10.866, 0.54464 10.839, 0.58779 10.809, 0.62932 10.777, 0.66913 10.743, 0.70711 10.707, 10.707 0.70711)