Geometry transformers

Geometry transformers

geos_centroid(geom)

geos_boundary(geom)

geos_minimum_width(geom)

geos_minimum_clearance_line(geom)

geos_minimum_rotated_rectangle(geom)

geos_minimum_bounding_circle(geom)

geos_unary_union(geom)

geos_coverage_union(geom)

geos_point_on_surface(geom)

geos_node(geom)

geos_make_valid(geom)

geos_unique_points(geom)

geos_reverse(geom)

geos_merge_lines(geom)

geos_build_area(geom)

geos_envelope(geom)

geos_convex_hull(geom)

geos_point_start(geom)

geos_point_end(geom)

geos_clone(geom)

geos_point_n(geom, index)

geos_simplify(geom, tolerance)

geos_simplify_preserve_topology(geom, tolerance)

geos_clip_by_rect(geom, rect)

Arguments

geom
A GEOS geometry vector

index
The index of the point or geometry to extract.

tolerance
A minimum distance to use for simplification. Use a higher value for more simplification.

rect
A list() representing rectangles in the form list(xmin, ymin, xmax, ymax). List items with length 1 will be recycled to the length of the longest item.

Value

A GEOS geometry vector of length geom

Examples

geos_centroid(c("POINT (0 1)", "LINESTRING (0 0, 1 1)"))

#> <geos_geometry[2]>
#> [1] POINT (0 1)     POINT (0.5 0.5)

geos_boundary(c("POLYGON ((0 0, 1 0, 0 1, 0 0))", "LINESTRING (0 0, 1 1)"))

#> <geos_geometry[2]>
#> [1] LINESTRING (0 0, 1 0, 0 1, 0 0) MULTIPOINT (0 0, 1 1)          

geos_minimum_width("POLYGON ((0 0, 1 0, 0 1, 0 0))")

#> <geos_geometry[1]>
#> [1] LINESTRING (0.5 0.5, 0 0)

geos_minimum_clearance_line("POLYGON ((0 0, 10 0, 10 10, 3 5, 0 10, 0 0))")

#> <geos_geometry[1]>
#> [1] LINESTRING (3 5, 0 5)

geos_minimum_rotated_rectangle("POLYGON ((0 0, 1 0, 0.5 0.5, 0 0))")

#> <geos_geometry[1]>
#> [1] POLYGON ((1 0, 1 0.5, 0 0.5, 0 0, 1 0))

geos_minimum_bounding_circle("LINESTRING (-1 -1, 1 1)")

#> <geos_geometry[1]>
#> [1] POLYGON ((1.4142 0, 1.387 -0.2759, 1.3066 -0.5412, 1.1759 -0.78569, 1 -1, 0.78569 -1.1759, 0.5412 -1.3066, 0.2759 -1.387, 2.2847e-15 -1.4142, -0.2759 -1.387, -0.5412 -1.3066, -0.78569 -1.1759, -1 -1, -1.1759 -0.78569, -1.3066 -0.5412, -1.387 -0.2759, -1.4142 -4.5695e-15, -1.387 0.2759, -1.3066 0.5412, -1.1759 0.78569, -1 1, -0.78569 1.1759, -0.5412 1.3066, -0.2759 1.387, -5.2841e-15 1.4142, 0.2759 1.387, 0.5412 1.3066, 0.78569 1.1759, 1 1, 1.1759 0.78569, 1.3066 0.5412, 1.387 0.2759, 1.4142 0))

geos_unary_union("MULTIPOINT (0 1, 0 1)")

#> <geos_geometry[1]>
#> [1] POINT (0 1)

geos_point_on_surface("LINESTRING (0 1, 0.2 3, 10 10)")

#> <geos_geometry[1]>
#> [1] POINT (0.2 3)

geos_node("POLYGON ((0 0, 1 0, 0 1, 0 0))")

#> <geos_geometry[1]>
#> [1] MULTILINESTRING ((0 0, 1 0, 0 1, 0 0))

geos_make_valid("POLYGON ((0 0, 1 1, 1 0, 0 1, 0 0))")

#> <geos_geometry[1]>
#> [1] MULTIPOLYGON (((0.5 0.5, 0 0, 0 1, 0.5 0.5)), ((0.5 0.5, 1 1, 1 0, 0.5 0.5)))

geos_unique_points("POLYGON ((0 0, 1 0, 0 1, 0 0))")

#> <geos_geometry[1]>
#> [1] MULTIPOINT (0 0, 1 0, 0 1)

geos_reverse("LINESTRING (0 0, 1 1)")

#> <geos_geometry[1]>
#> [1] LINESTRING (1 1, 0 0)

geos_merge_lines(
  "MULTILINESTRING ((0 0, 0.5 0.5, 2 2), (0.5 0.5, 2 2))"
)

#> <geos_geometry[1]>
#> [1] LINESTRING (0 0, 0.5 0.5, 2 2, 0.5 0.5)

geos_build_area("LINESTRING (0 0, 1 0, 0 1, 0 0)")

#> <geos_geometry[1]>
#> [1] POLYGON ((0 0, 0 1, 1 0, 0 0))

geos_envelope("LINESTRING (0 0, 1 2)")

#> <geos_geometry[1]>
#> [1] POLYGON ((0 0, 1 0, 1 2, 0 2, 0 0))

geos_convex_hull("MULTIPOINT (0 0, 1 0, 0 2, 0 0)")

#> <geos_geometry[1]>
#> [1] POLYGON ((0 0, 0 2, 1 0, 0 0))

geos_point_start("LINESTRING (0 0, 1 1)")

#> <geos_geometry[1]>
#> [1] POINT (0 0)

geos_point_end("LINESTRING (0 0, 1 1)")

#> <geos_geometry[1]>
#> [1] POINT (1 1)

geos_interpolate("LINESTRING (0 0, 1 1)", 1)

#> <geos_geometry[1]>
#> [1] POINT (0.70711 0.70711)

geos_interpolate_normalized("LINESTRING (0 0, 1 1)", 1)

#> <geos_geometry[1]>
#> [1] POINT (1 1)

geos_simplify("LINESTRING (0 0, 0 1, 0 2)", 0.1)

#> <geos_geometry[1]>
#> [1] LINESTRING (0 0, 0 2)

geos_simplify_preserve_topology("LINESTRING (0 0, 0 1, 0 2)", 0.1)

#> <geos_geometry[1]>
#> [1] LINESTRING (0 0, 0 2)