Read and write well-known text

Read and write well-known text

geos_read_wkt(wkt)

geos_write_wkt(geom, include_z = TRUE, precision = 16, trim = TRUE)

geos_read_wkb(wkb)

geos_write_wkb(geom, include_z = TRUE, include_srid = FALSE, endian = 1)

geos_read_hex(hex)

geos_write_hex(geom, include_z = TRUE, include_srid = FALSE, endian = 1)

geos_read_xy(point)

geos_write_xy(geom)

Arguments

wkt
A character vector containing well-known text.

geom
A GEOS geometry vector

include_z
Include the values of the Z and M coordinates and/or SRID in the output? Use FALSE to omit, TRUE to include, or NA to include only if present. Note that using TRUE may result in an error if there is no value present in the original.

precision
The rounding precision to use when writing (number of decimal places).

trim
Trim unnecessary zeroes in the output?

wkb
A list() of raw() vectors, such as that returned by sf::st_as_binary().

include_srid
Include the values of the Z and M coordinates and/or SRID in the output? Use FALSE to omit, TRUE to include, or NA to include only if present. Note that using TRUE may result in an error if there is no value present in the original.

endian
For WKB writing, 0 for big endian, 1 for little endian. Defaults to wk_platform_endian() (slightly faster).

hex
A hexidecimal representation of well-known binary

point
A list() representing points in the form list(x, y).

Examples

geos_read_wkt("POINT (30 10)")

#> <geos_geometry[1]>
#> [1] <POINT (30 10)>

geos_write_wkt(geos_read_wkt("POINT (30 10)"))

#> [1] "POINT (30 10)"